renamed react exercise folders to be camelCase consistently
