+++ frontend/src/exercises/colourfulSemantics/ColourfulSemanticsBoard.tsx
