+++ frontend/src/exercises/colourfulSemantics/ColourfulSemanticsGame.tsx
