+++ frontend/src/exercises/colourfulSemantics/boardUtils.ts
