+++ frontend/src/exercises/colourfulSemantics/configureScene.ts
