+++ frontend/src/exercises/colourfulSemantics/data/advancedVariant.json
