+++ frontend/src/exercises/colourfulSemantics/data/earlyYearsVariant.json
