+++ frontend/src/exercises/colourfulSemantics/data/sharedAssetPool.json
