+++ frontend/src/exercises/colourfulSemantics/data/standardVariant.json
