+++ frontend/src/exercises/colourfulSemantics/slotMetadata.ts
