+++ frontend/src/exercises/colourfulSemantics/types.ts
