+++ frontend/src/exercises/colourfulSemantics/variantData.ts
