+++ frontend/src/exercises/inTheKnow/InTheKnow.tsx
