+++ frontend/src/exercises/inTheKnow/InTheKnowGame.tsx
