+++ frontend/src/exercises/inTheKnow/exerciseData.json
