+++ frontend/src/exercises/inTheKnow/types.ts
