+++ frontend/src/exercises/storyTrain/StoryTrain.tsx
