+++ frontend/src/exercises/storyTrain/StoryTrainBoard.module.css
