+++ frontend/src/exercises/storyTrain/StoryTrainBoard.tsx
