+++ frontend/src/exercises/storyTrain/StoryTrainGame.tsx
