+++ frontend/src/exercises/storyTrain/boardUtils.ts
