+++ frontend/src/exercises/storyTrain/exerciseData.advanced.json
