+++ frontend/src/exercises/storyTrain/types.ts
