+++ frontend/src/exercises/whatHappensNext/WhatHappensNext.tsx
