+++ frontend/src/exercises/whatHappensNext/WhatHappensNextGame.tsx
