+++ frontend/src/exercises/whatHappensNext/exerciseData.json
