+++ frontend/src/exercises/whatHappensNext/types.ts
