+++ frontend/src/exercises/whatHappensNext/whatHappensNext.module.css
