+++ frontend/src/exercises/whatsInTheBag/WhatsInTheBag.tsx
