+++ frontend/src/exercises/whatsInTheBag/WhatsInTheBagGame.tsx
