+++ frontend/src/exercises/whatsInTheBag/WhatsInTheBagSettingsScreen.tsx
