+++ frontend/src/exercises/whatsInTheBag/exerciseData.json
