+++ frontend/src/exercises/whatsInTheBag/types.ts
