+++ static/js/exercises/css/concept-quest.css
@@ -1,1 +1,1 @@
-._settingsPanel_xrdmn_1{display:flex;flex-direction:column;gap:1.5rem}._settingsGroup_xrdmn_7{display:flex;flex-direction:column;gap:.85rem}._settingsHeading_xrdmn_13{margin:0;font-size:1.125rem;font-weight:700}._settingsHint_xrdmn_19{margin:0;color:#4b5563;font-size:.95rem;line-height:1.5}._selectionGrid_xrdmn_26{display:grid;grid-template-columns:repeat(2,minmax(0,1fr));gap:.75rem}._selectionGridCompact_xrdmn_32{display:grid;grid-template-columns:repeat(3,minmax(0,1fr));gap:.75rem}._optionsGrid_xrdmn_38{display:flex;flex-wrap:wrap;justify-content:center;gap:1.25rem;width:100%;max-width:1200px;margin:0 auto}._optionsGrid_xrdmn_38 button{width:clamp(180px,24vw,280px);height:clamp(180px,24vw,280px);padding:1.25rem}._optionsGrid_xrdmn_38 img{width:clamp(130px,20vw,220px);height:clamp(130px,20vw,220px)}@media(max-width:640px){._selectionGrid_xrdmn_26,._selectionGridCompact_xrdmn_32{grid-template-columns:repeat(2,minmax(0,1fr))}._optionsGrid_xrdmn_38{gap:.75rem}._optionsGrid_xrdmn_38 button{width:clamp(130px,40vw,180px);height:clamp(130px,40vw,180px)}._optionsGrid_xrdmn_38 img{width:clamp(90px,28vw,130px);height:clamp(90px,28vw,130px)}}
+._container_1tz2v_1{display:flex;flex-direction:column;gap:1.25rem}._section_1tz2v_7{display:flex;flex-direction:column;gap:.5rem}._sectionLabel_1tz2v_13{margin:0;font-size:.95rem;font-weight:600;color:#2c233f}._complexityToggleRow_1tz2v_20{display:flex;align-items:center;gap:.45rem;font-size:.95rem;color:#4b5563}._complexityToggleRow_1tz2v_20 input{margin:0}._categoriesGrid_1tz2v_32{display:grid;grid-template-columns:repeat(auto-fit,minmax(110px,1fr));gap:.65rem}._categoryCard_1tz2v_38{position:relative;display:flex;align-items:center;justify-content:center;padding:1rem .75rem;border:2px solid #d0d7de;border-radius:.75rem;background-color:#fff;cursor:pointer;transition:all .2s ease;min-height:80px;text-align:center;font-weight:500;color:#2c233f;font-size:.85rem}._categoryCard_1tz2v_38:hover{border-color:#5b8def;background-color:#f0f4ff;transform:translateY(-1px)}._categoryCard_1tz2v_38._selected_1tz2v_62{border-color:#33da73;background-color:#33da731a;color:#2c233f}._categoryCard_1tz2v_38._selected_1tz2v_62:after{content:"✓";position:absolute;top:.25rem;right:.25rem;width:1.25rem;height:1.25rem;background-color:#33da73;color:#fff;border-radius:50%;display:flex;align-items:center;justify-content:center;font-weight:700;font-size:.65rem}._optionsCard_1tz2v_85{display:flex;flex-direction:column;gap:.5rem;padding:1rem;border:2px solid #d0d7de;border-radius:.75rem;background-color:#fff}._cardTitle_1tz2v_95{display:inline;font-weight:500}._cardReference_1tz2v_100{display:inline;font-size:.92rem;color:#4b5563;font-weight:500}._optionsGrid_1tz2v_107{display:flex;flex-wrap:wrap;justify-content:center;gap:1.25rem;width:100%;max-width:1200px;margin:0 auto}._optionsGrid_1tz2v_107 button{width:clamp(180px,24vw,280px);height:clamp(180px,24vw,280px);padding:1.25rem}._optionsGrid_1tz2v_107 img{width:clamp(130px,20vw,220px);height:clamp(130px,20vw,220px)}@media(max-width:640px){._container_1tz2v_1{gap:.8rem}._categoriesGrid_1tz2v_32{grid-template-columns:repeat(2,1fr);gap:.4rem}._categoryCard_1tz2v_38{min-height:65px;padding:.6rem .4rem;font-size:.75rem}._optionsCard_1tz2v_85{padding:.65rem}._sectionLabel_1tz2v_13,._complexityToggleRow_1tz2v_20{font-size:.85rem}._optionsGrid_1tz2v_107{gap:.75rem}._optionsGrid_1tz2v_107 button{width:clamp(130px,40vw,180px);height:clamp(130px,40vw,180px)}._optionsGrid_1tz2v_107 img{width:clamp(90px,28vw,130px);height:clamp(90px,28vw,130px)}}
